Use float32 for JAX time and observation arrays
float64 was requested but silently truncated to float32 (JAX_ENABLE_X64
not set). The extra precision is unnecessary — time values are fractional
days (small numbers) and observations are typical sensor/survey data.

+++ src/causal_ssm_agent/flows/stages/stage4b_parametric_id.py
@@ -63,8 +63,8 @@ def parametric_id_task(
         ssm_model = builder._model
 
         # Extract observations and times
-        observations = jnp.array(X.drop("time").to_numpy(), dtype=jnp.float64)
-        times = jnp.array(X["time"].to_numpy(), dtype=jnp.float64)
+        observations = jnp.array(X.drop("time").to_numpy(), dtype=jnp.float32)
+        times = jnp.array(X["time"].to_numpy(), dtype=jnp.float32)
         T = int(times.shape[0])
 
         # T-rule: fast necessary condition (hard gate)
